moved lifetime tracker codes to helper package
