--- carpetmodSrc/carpet/commands/lifetime/removal/RemovalReason.java
@@ -1,7 +1,0 @@
-package carpet.commands.lifetime.removal;
-
-import carpet.commands.lifetime.utils.AbstractReason;
-
-public abstract class RemovalReason extends AbstractReason
-{
-}
