--- carpetmodSrc/carpet/commands/lifetime/spawning/SpawningReason.java
@@ -1,7 +1,0 @@
-package carpet.commands.lifetime.spawning;
-
-import carpet.commands.lifetime.utils.AbstractReason;
-
-public abstract class SpawningReason extends AbstractReason
-{
-}
