+++ carpetmodSrc/carpet/helpers/lifetime/removal/RemovalReason.java
@@ -1,0 +1,7 @@
+package carpet.helpers.lifetime.removal;
+
+import carpet.helpers.lifetime.utils.AbstractReason;
+
+public abstract class RemovalReason extends AbstractReason
+{
+}
