+++ carpetmodSrc/carpet/helpers/lifetime/spawning/SpawningReason.java
@@ -1,0 +1,7 @@
+package carpet.helpers.lifetime.spawning;
+
+import carpet.helpers.lifetime.utils.AbstractReason;
+
+public abstract class SpawningReason extends AbstractReason
+{
+}
