+++ carpetmodSrc/carpet/helpers/lifetime/trackeddata/ExperienceOrbTrackedData.java
@@ -1,4 +1,4 @@
-package carpet.commands.lifetime.trackeddata;
+package carpet.helpers.lifetime.trackeddata;
 
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.item.EntityXPOrb;
