+++ carpetmodSrc/carpet/helpers/lifetime/trackeddata/ExtraCountTrackedData.java
@@ -1,9 +1,9 @@
-package carpet.commands.lifetime.trackeddata;
+package carpet.helpers.lifetime.trackeddata;
 
-import carpet.commands.lifetime.removal.RemovalReason;
-import carpet.commands.lifetime.spawning.SpawningReason;
-import carpet.commands.lifetime.utils.CounterUtil;
-import carpet.commands.lifetime.utils.TextUtil;
+import carpet.helpers.lifetime.removal.RemovalReason;
+import carpet.helpers.lifetime.spawning.SpawningReason;
+import carpet.helpers.lifetime.utils.CounterUtil;
+import carpet.helpers.lifetime.utils.TextUtil;
 import carpet.utils.Messenger;
 import com.google.common.collect.Maps;
 import net.minecraft.entity.Entity;
