+++ carpetmodSrc/carpet/helpers/lifetime/utils/LifeTimeStatistic.java
@@ -1,4 +1,4 @@
-package carpet.commands.lifetime.utils;
+package carpet.helpers.lifetime.utils;
 
 import carpet.utils.Messenger;
 import net.minecraft.entity.Entity;
