+++ carpetmodSrc/carpet/helpers/lifetime/utils/LifeTimeTrackerUtil.java
@@ -1,4 +1,4 @@
-package carpet.commands.lifetime.utils;
+package carpet.helpers.lifetime.utils;
 
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityList;
