+++ carpetmodSrc/carpet/helpers/lifetime/utils/SpecificDetailMode.java
@@ -1,4 +1,4 @@
-package carpet.commands.lifetime.utils;
+package carpet.helpers.lifetime.utils;
 
 import java.util.Arrays;
