+++ carpetmodSrc/carpet/helpers/lifetime/utils/TextUtil.java
@@ -1,4 +1,4 @@
-package carpet.commands.lifetime.utils;
+package carpet.helpers.lifetime.utils;
 
 import carpet.utils.Messenger;
 import com.google.common.collect.Maps;
