+++ patches/net/minecraft/block/BlockPortal.java.patch
@@ -3,7 +3,7 @@
 @@ -1,5 +1,6 @@
  package net.minecraft.block;
  
-+import carpet.commands.lifetime.spawning.LiteralSpawningReason;
++import carpet.helpers.lifetime.spawning.LiteralSpawningReason;
  import com.google.common.cache.LoadingCache;
  import java.util.Random;
  import javax.annotation.Nullable;
