+++ patches/net/minecraft/entity/EntityLivingBase.java.patch
@@ -3,7 +3,7 @@
 @@ -1,5 +1,6 @@
  package net.minecraft.entity;
  
-+import carpet.commands.lifetime.spawning.MobDropSpawningReason;
++import carpet.helpers.lifetime.spawning.MobDropSpawningReason;
  import com.google.common.base.Objects;
  import com.google.common.collect.Maps;
  import java.util.Collection;
