+++ patches/net/minecraft/entity/monster/EntityZombie.java.patch
@@ -5,7 +5,7 @@
  import java.util.UUID;
  import javax.annotation.Nullable;
 +
-+import carpet.commands.lifetime.spawning.LiteralSpawningReason;
++import carpet.helpers.lifetime.spawning.LiteralSpawningReason;
  import net.minecraft.block.Block;
  import net.minecraft.entity.Entity;
  import net.minecraft.entity.EntityLiving;
