+++ patches/net/minecraft/world/WorldEntitySpawner.java.patch
@@ -3,7 +3,7 @@
 @@ -1,5 +1,6 @@
  package net.minecraft.world;
  
-+import carpet.commands.lifetime.spawning.LiteralSpawningReason;
++import carpet.helpers.lifetime.spawning.LiteralSpawningReason;
  import com.google.common.collect.Sets;
  import java.util.List;
  import java.util.Random;
