+++ patches/net/minecraft/world/chunk/storage/AnvilChunkLoader.java.patch
@@ -4,7 +4,7 @@
  package net.minecraft.world.chunk.storage;
  
 -import com.google.common.collect.Maps;
-+import carpet.commands.lifetime.spawning.LiteralSpawningReason;
++import carpet.helpers.lifetime.spawning.LiteralSpawningReason;
 +import com.google.common.collect.Maps; //CM unused import
  import java.io.DataInputStream;
  import java.io.DataOutputStream;
